Universe Visual Update
Click on a squib to select it. All currently selected squibs can be
moved around. For now, click on the red rectangle to deselect all
squibs. When no squibs are selected, everything is moved together.
Still need to change how the squibs move and will add some more
functionality to it.

+++ src/application/VcPtVisualSchematicViewClickable.java
@@ -29,8 +29,8 @@ public class VcPtVisualSchematicViewClickable extends VcPtVisualSchematicView {
 		int x, y, xt;
 		
 		// Positions to start drawing
-        x = 50 + mouseInfo.offX();
-        y = 50 + mouseInfo.offY();
+        x = 50 + (int)mouseInfo.offX();
+        y = 50 + (int)mouseInfo.offY();
         xt = x;
         		
         // sets the start of a mouse drag
